Working Game
Did a lot to this game. I added a "New Game" button. It has quirks
at the very beginning where the calculated sum may be 0+0+0=1.
That would need to be handled so that the initial value is zero.
I wonder how the possible solution did it. I must have fiddled with
this exercise for a very long long time. And yet, I didn't look at
the solution. this is what I want from my students. But how do we
get ther? I am not sure.